refactor(agent): share one base64 codec via shared/base64
The voice Durable Object hand-rolled `bytesToBase64`/`base64ToBytes`, a
byte-for-byte copy of the chunked `btoa`/`atob` helpers already private to
`shared/wire-codec.ts`. Extract both into a new bundler-inlined
`shared/base64.ts` and have the wire codec (`@lunora/client` + `@lunora/do`)
and the voice DO (`@lunora/agent`) import the single implementation.

`@lunora/agent` now consumes `shared/*`, so its tsconfig drops the vestigial
`outDir`/`rootDir` (a set `rootDir` raises TS6059 for the inlined out-of-package
file) with the same breadcrumb the other shared consumers carry.
