+++ packages/agent/tsconfig.json
@@ -1,8 +1,11 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": {
-        "outDir": "dist",
-        "rootDir": ".",
+        // Intentionally no `outDir`/`rootDir` (unlike sibling packages): `lint:types`
+        // is `tsc --noEmit` and the real build is packem (driven by package.json
+        // exports), so neither is load-bearing here. A set `rootDir` would raise
+        // TS6059 for the inlined `../../../shared/base64` import (a file outside
+        // this package, bundled in at build time). See shared/base64.ts.
         "moduleResolution": "bundler",
         "types": ["node"]
     },
